api/requirements.txt
google-cloud-aiplatform==1.97.0
Flask
google-cloud-logging
